refactor: Add image service to Fan.Blog #266
