+++ src/Fan.Blog/Services/BlogServiceConfig.cs
@@ -11,11 +11,6 @@ namespace Fan.Blog.Services
     /// </summary>
     public partial class BlogService
     {
-        /// <summary>
-        /// "Blog"
-        /// </summary>
-        public const string BLOG_APP_NAME = "Blog";
-
         // -------------------------------------------------------------------- Posts
 
         /// <summary>
@@ -28,103 +23,5 @@ namespace Fan.Blog.Services
         /// How many words to extract into excerpt from body. Default 55.
         /// </summary>
         public const int EXCERPT_WORD_LIMIT = 55;
-
-        // -------------------------------------------------------------------- Images
-
-        /// <summary>
-        /// Blog accepted image types: .jpg .jpeg .png .gif
-        /// </summary>
-        /// <remarks>
-        /// Got the idea from WP https://en.support.wordpress.com/images/
-        /// For accepted file types https://en.support.wordpress.com/accepted-filetypes/
-        /// </remarks>
-        public static readonly string[] Accepted_Image_Types = { ".jpg", ".jpeg", ".gif", ".png" };
-
-        /// <summary>
-        /// The separator used in image paths is '/'.
-        /// </summary>
-        /// <remarks>
-        /// All <see cref="IStorageProvider"/> implementations should take this separator and replace 
-        /// it with your specific one.
-        /// </remarks>
-        public const char IMAGE_PATH_SEPARATOR = '/';
-
-        /// <summary>
-        /// Large image size 1200 pixel.
-        /// </summary>
-        public const int LARGE_IMG_SIZE = 1200;
-
-        /// <summary>
-        /// Medium image size 800 pixel.
-        /// </summary>
-        public const int MEDIUM_IMG_SIZE = 800;
-
-        /// <summary>
-        /// Small image size 400 pixel.
-        /// </summary>
-        public const int SMALL_IMG_SIZE = 400;
-
-        /// <summary>
-        /// The different image resizes per image upload.
-        /// </summary>
-        /// <param name="uploadedOn"></param>
-        /// <returns></returns>
-        public static List<ImageResizeInfo> GetImageResizeList(DateTimeOffset uploadedOn)
-        {
-            return new List<ImageResizeInfo> {
-                new ImageResizeInfo {
-                    TargetSize = int.MaxValue,
-                    Path = GetImagePath(uploadedOn, EImageSize.Original),
-                    PathSeparator = IMAGE_PATH_SEPARATOR,
-                },
-                new ImageResizeInfo {
-                    TargetSize = LARGE_IMG_SIZE,
-                    Path = GetImagePath(uploadedOn, EImageSize.Large),
-                    PathSeparator = IMAGE_PATH_SEPARATOR,
-                },
-                new ImageResizeInfo {
-                    TargetSize = MEDIUM_IMG_SIZE,
-                    Path = GetImagePath(uploadedOn, EImageSize.Medium),
-                    PathSeparator = IMAGE_PATH_SEPARATOR,
-                },
-                new ImageResizeInfo {
-                    TargetSize = SMALL_IMG_SIZE,
-                    Path = GetImagePath(uploadedOn, EImageSize.Small),
-                    PathSeparator = IMAGE_PATH_SEPARATOR,
-                },
-            };
-        }
-
-        /// <summary>
-        /// Returns the stored image path, "{app}/{year}/{month}" or "{app}/{year}/{month}/{sizePath}".
-        /// </summary>
-        /// <param name="uploadedOn"></param>
-        /// <param name="size"></param>
-        /// <returns></returns>
-        public static string GetImagePath(DateTimeOffset uploadedOn, EImageSize size)
-        {
-            var app = BLOG_APP_NAME.ToLowerInvariant();
-            var year = uploadedOn.Year.ToString();
-            var month = uploadedOn.Month.ToString("d2");
-            var sizePath = "";
-
-            switch (size)
-            {
-                case EImageSize.Large:
-                    sizePath = "lg";
-                    break;
-                case EImageSize.Medium:
-                    sizePath = "md";
-                    break;
-                case EImageSize.Small:
-                    sizePath = "sm";
-                    break;
-                default:
-                    sizePath = null;
-                    break;
-            }
-
-            return size == EImageSize.Original ? $"{app}/{year}/{month}" : $"{app}/{year}/{month}/{sizePath}";
-        }
     }
 }
